use int64 in interface
Also remove the "debug" client and remove the tests that import
"statsd-interface" since we are no longer compliant with that
interface.
